[DB] Fix builtin DB init script. Document the DB existingSecret  params
In case of Builtin DBs:
Init script was giving ownership of all DBs to georchestra user, no using
the config params for geodata and datafeeder DBs. Which were nevertheless
used by the geodata and datafeeder connection secrets. So for instance
datafeeder was trying to connect with `datafeeder` user, but it had not
been created.
This way, the init script actually creates the users and gives them
ownership on their respective DB

Documentation: clarify the usage of the existingSecret params