geoserver - adding fqdn to the geoserver's CSRF whitelist
See https://docs.geoserver.org/main/en/user/security/webadmin/csrf.html
for the motivation.

Also please note that in case of using the geOrchestra gateway, the
default http header for the referrer-policy will be "no-referrer", which
prevents wicket to determine if the request is legit or not (the check
is based on the referer or origin http headers). so to make sure the
referer header is provided, this requires the following configuration
property to be set on the gateway as well:

```
spring.cloud.gateway.filter.secure-headers=strict-origin (or something
  laxer than no-referrer, see mdn doc for possible values)
```